+++ src/components/notes/table/NoteTable.js
@@ -110,7 +110,9 @@ function NoteTable() {
     });
 
     return (
-        <div style={{ overflowY: 'hidden', height: 'calc(100% - 40px)' }}>
+        <div
+            className="joyride-note-table"
+            style={{ overflowY: 'hidden', height: 'calc(100% - 40px)' }}>
             <AutoSizer>
                 {({ height }) => (
                     <Table