+++ src/components/common/HelpButton.js
@@ -1,16 +1,16 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Icon from 'components/common/Icon';
-import withJoyride from 'containers/WithJoyride';
 import { useJoyrideApi } from 'hooks/UseJoyrideApi';
 
-function HelpButton(props) {
+function HelpButton({ id }) {
     const joyrideApi = useJoyrideApi();
 
     const onClick = () => {
         joyrideApi.setJoyrideOptions({
-            id: props.id,
-            run: true
+            id,
+            run: true,
+            stepIndex: 0
         });
     };
 
@@ -29,4 +29,4 @@ HelpButton.propTypes = {
     id: PropTypes.string.isRequired
 };
 
-export default withJoyride(HelpButton);
+export default HelpButton;