+++ src/containers/WithJoyride.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Joyride, { STATUS } from 'react-joyride';
+import Joyride, { ACTIONS, EVENTS, STATUS } from 'react-joyride';
 import { getConfig } from 'constants/JoyrideConfig';
 import { useJoyrideApi } from 'hooks/UseJoyrideApi';
 
@@ -8,9 +8,13 @@ function withJoyride(Component) {
         const joyrideApi = useJoyrideApi();
 
         const callback = data => {
-            const { status } = data;
+            const { action, index, status, type } = data;
 
-            if ([STATUS.FINISHED, STATUS.SKIPPED].includes(status)) {
+            if ([EVENTS.STEP_AFTER, EVENTS.TARGET_NOT_FOUND].includes(type)) {
+                joyrideApi.setJoyrideOptions({ stepIndex: index + (action === ACTIONS.PREV ? -1 : 1) });
+            } else if ([ACTIONS.CLOSE].includes(action)) {
+                joyrideApi.setJoyrideOptions({ run: false });
+            } else if ([STATUS.FINISHED, STATUS.SKIPPED].includes(status)) {
                 joyrideApi.setJoyrideOptions({ run: false });
             }
         };
@@ -19,10 +23,20 @@ function withJoyride(Component) {
             <React.Fragment>
                 <Joyride
                     run={!!joyrideApi.joyride.run}
+                    stepIndex={joyrideApi.joyride.stepIndex}
                     callback={callback}
                     styles={{
                         options: {
                             zIndex: 2000
+                        },
+                        tooltip: {
+                            fontSize: 13
+                        },
+                        tooltipContainer: {
+                            textAlign: 'left'
+                        },
+                        tooltipTitle: {
+                            fontSize: 15
                         }
                     }}
                     {...getConfig(joyrideApi.joyride.id)} />